#143 Add scalastyle and fix InteliJ warnings (#145)
* Added scalastyle to maven with default spark configuration

* Customised scalacheck rules

* scalastyle conformance for sparkle

* Scalachecked: api, cli

* Limited scalastyle passing on main sources

* Updated scala-maven-plugin

* Fixed most of InteliJ inspection warnigns

* Added InteliJ inspection profile

* Added into on scalastyle and InteliJ to CONTRIBUTING.md

+++ src/main/scala/au/csiro/variantspark/input/HashingLabelSource.scala
@@ -2,5 +2,5 @@ package au.csiro.variantspark.input
 
 class HashingLabelSource(val maxLabel: Int) extends LabelSource {
   def getLabels(labels: Seq[String]): Array[Int] =
-    labels.map(l => (math.abs(l.hashCode()) % maxLabel).toInt).toArray
+    labels.map(l => math.abs(l.hashCode()) % maxLabel).toArray
 }
